+++ sdmx/reader/sdmxml.py
@@ -12,6 +12,7 @@ from lxml import etree
 from lxml.etree import QName, XPath
 
 from sdmx.exceptions import ParseError, XMLParseError
+from sdmx.format.xml import NS, qname
 from sdmx.message import (
     DataMessage, ErrorMessage, Footer, Header, StructureMessage,
     )
@@ -43,25 +44,6 @@ URN = re.compile(r'urn:sdmx:org\.sdmx\.infomodel'
                  r'\.(?P<class>[^=]*)=((?P<agency>[^:]*):)?'
                  r'(?P<id>[^\(\.]*)(\((?P<version>[\d\.]*)\))?'
                  r'(\.(?P<item_id>.*))?')
-
-
-# XML namespaces
-_base_ns = 'http://www.sdmx.org/resources/sdmxml/schemas/v2_1'
-NS = {
-    'com': f'{_base_ns}/common',
-    'data': f'{_base_ns}/data/structurespecific',
-    'str': f'{_base_ns}/structure',
-    'mes': f'{_base_ns}/message',
-    'gen': f'{_base_ns}/data/generic',
-    'footer': f'{_base_ns}/message/footer',
-    'xml': 'http://www.w3.org/XML/1998/namespace',
-    'xsi': 'http://www.w3.org/2001/XMLSchema-instance',
-    }
-
-
-def qname(ns, name):
-    """Return a fully-qualified tag *name* in namespace *ns*."""
-    return QName(NS[ns], name)
 
 
 _TO_SNAKE_RE = re.compile('([A-Z]+)')
